requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
web3==6.11.1
python-dotenv==1.0.0
requests==2.31.0